move test cases of common behaviour
move test cases of common behaviour from InputDateTimePicker to createInputPicker
