+++ packages/components/src/DateTimePickers/shared/createInputPicker.test.js
@@ -1,5 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
+import keycode from 'keycode';
+
 import createInputPicker from './createInputPicker';
 import { DateTimeContext } from '../DateTime/Context';
 
@@ -79,6 +81,41 @@ describe('InputPicker', () => {
 
 			// then
 			expect(onBlur).toBeCalled();
+		});
+	});
+
+	describe('keydown', () => {
+		it('should close the picker and focus on input with ESC', () => {
+			// given
+			const wrapper = mount(
+				<DateTimeContext.Provider value={managerValue}>
+					<InputSomethingPicker id="my-id" />
+				</DateTimeContext.Provider>);
+			wrapper.simulate('focus');
+			expect(getOverlay(wrapper).exists()).toBe(true);
+			const event = { keyCode: keycode.codes.esc };
+
+			// when
+			wrapper.simulate('keydown', event);
+
+			// then
+			expect(getOverlay(wrapper).exists()).toBe(false);
+		});
+
+		it('should open picker if it is closed with DOWN on input', () => {
+			// given
+			const wrapper = mount(
+				<DateTimeContext.Provider value={managerValue}>
+					<InputSomethingPicker id="my-id" />
+				</DateTimeContext.Provider>);
+			expect(getOverlay(wrapper).exists()).toBe(false);
+			const event = { keyCode: keycode.codes.down };
+
+			// when
+			wrapper.find('input').simulate('keydown', event);
+
+			// then
+			expect(getOverlay(wrapper).exists()).toBe(true);
 		});
 	});
